Smart class was created
